backend/requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.38.0
python-dotenv==1.2.1
SQLAlchemy==2.0.44
pydantic==2.12.3
email-validator==2.3.0
python-multipart==0.0.20
firebase-admin==7.1.0
openai==2.6.1
pypdf==6.1.3
requests==2.32.5
certifi==2025.10.5
psycopg[binary]==3.2.12